refactor(screen-engine): introduce renderer registry runtime
